# Optimize Pascal implementations and add Delphi benchmarks