--- a/fibonacci/delphi/README.md
+++ b/fibonacci/delphi/README.md
@@ -1,0 +1,27 @@
+# Delphi Fibonacci Implementation
+
+This implementation requires Delphi to compile. Tested with Delphi 2007.
+
+## Performance
+On test hardware:
+- Original: Not previously implemented
+- Current: 871.8ms
+- Compared to other languages on same hardware:
+  - Kotlin: 19.0ms
+  - Java: 138.6ms
+  - Node: 1.505s
+  - Python: 27.318s
+
+## Compiler Settings Used
+- Build Configuration: Release
+- Optimization: Enabled
+- Stack frames: Enabled
+- Range checking: Disabled
+- I/O checking: Disabled
+- Overflow checking: Disabled
+
+## Implementation Notes
+- Uses same recursive algorithm as other implementations
+- Optimized using Cardinal (unsigned 32-bit) integers
+- Optimized string parsing with Val function
+- Windows-only due to Delphi platform limitations