--- a/run.sh
+++ b/run.sh
@@ -49,3 +49,4 @@ run "C#" "./csharp/code/csharp 40"
 run "Ruby" "ruby ./ruby/code.rb 40"
 run "Ruby YJIT" "miniruby --yjit ./ruby/code.rb 40"
 run "Haskell" "./hs/code 40"
+run "Delphi" "./delphi/code.exe 40" # Delphi compilation requires Delphi 2007 or later IDE