- Added Delphi implementations for comparison:
  * Fibonacci: 871.8ms
  * Loops: 1.719s

- Added optimized Free Pascal implementations:
  * Fibonacci: ~15.6% faster (1.039s -> 876.7ms)
  * Loops: Similar performance with better type safety


Changes include:
- Changed to Cardinal type for unsigned integers
- Added compiler optimization directives
- Optimized parameter parsing
- Added project files with optimal compiler settings
- Added documentation with benchmarks and settings

Performance measured on Windows with both Delphi 2007 and FPC.