@oldfort76 2024-12-05
Updated the compile script for free pascal (using the optimizations that are always used in an actual release: there's no point in sending the debug version for production). And updated the run script to run the delphi exe. I suppose it's hard to duplicate the results since the user will need to compile himself (as I cant upload a exe here)... But if someone has Delphi, even if a newer version, it will work.